dict.c: Free non-strduped av_dict_set arguments on error.
dict.c: Free non-strduped av_dict_set arguments on error.

Unfortunately this was not explicitly documented and thus
might be risky.
But all uses I could find in FFmpeg and one in VLC had a memleak
in these cases, and I could not find any that relied on the previous
behaviour.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>